Readd missing HoodIO setting to RobotContainer, move IntakePivotSystem to new folder, and create empty files for Akit refactor.
