+++ src/main/java/org/team157/robot/subsystems/intakePivot/IntakePivot.java
@@ -1,0 +1,7 @@
+package org.team157.robot.subsystems.intakePivot;
+
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+
+public class IntakePivot extends SubsystemBase {
+    
+}
