+++ src/main/java/org/team157/robot/subsystems/intakePivot/IntakePivotIO.java
@@ -1,0 +1,5 @@
+package org.team157.robot.subsystems.intakePivot;
+
+public interface IntakePivotIO {
+    
+}
